ks und ov Basisvariablentypen auf 32 Bit festgenagelt (uint32_t). Grob unter windows getestet.
git-svn-id: https://dev.plt.rwth-aachen.de/acplt-repo@8720 7a3915f3-eddd-4ae9-ab37-eb2ea51f0265

+++ history/archive/acplt/trunk/ks/include/ks/ks.h
@@ -34,6 +34,8 @@
 
 #include "ks/rpc.h"
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -294,7 +296,7 @@ ENUMDEF(KS_STRUCTURE_FLAGS)
  * instance, function block system servers might want to indicate that par-
  * ticular variables are input or output ports of function blocks, etc.
  */
-typedef u_long KS_SEMANTIC_FLAGS;
+typedef uint32_t KS_SEMANTIC_FLAGS;
 
 
 /* ----------------------------------------------------------------------------
